+++ semester_3/lecture_16_advanced_ml/neural_network/Algorithm.java
@@ -46,7 +46,7 @@ public class Algorithm {
         System.out.println("=".repeat(70));
         
         Algorithm algo = Algorithm.create();
-        float result = algo.sigmoid(null);
+        int result = algo.sigmoid(null);
         System.out.println("Result: " + result);
         System.out.println("=".repeat(70));
     }
