+++ semester_5/lecture_27_hyperparameter_optimization/bayesian_optimization/Algorithm.java
@@ -47,7 +47,7 @@ public class Algorithm {
         System.out.println("=".repeat(70));
         
         Algorithm algo = Algorithm.create();
-        float result = algo._acquisition_function("", null);
+        int result = algo._acquisition_function("", null);
         System.out.println("Result: " + result);
         System.out.println("=".repeat(70));
     }
